requirements.txt
#API
fastapi
uvicorn
httpx
pytest-asyncio
pydantic
imageio


#GCP
google-cloud-storage


#YOLO
ultralytics
